requirements.txt
boto3==1.26.142
pandas==1.5.3
pip==22.3.1
python-dotenv==1.0.0
